Change type to int16_t

+++ evalAssignment1/update_locations.c
@@ -1,20 +1,20 @@
 #include "update_locations.h"
 
-// Create a list of 'size' int8_ting point numbers in the range [-bound, bound]
-int8_t * generate_random_list(int32_t size, int32_t bound) {
-    int8_t* array;
-    array = (int8_t*)malloc(sizeof(int8_t) * size);
+// Create a list of 'size' int16_ting point numbers in the range [-bound, bound]
+int16_t * generate_random_list(int32_t size, int32_t bound) {
+    int16_t* array;
+    array = (int16_t*)malloc(sizeof(int16_t) * size);
 
 
     for (int i = 0; i < size; i++) {
-        int8_t num = (int8_t)(rand() % (2*bound+1)) - bound;
+        int16_t num = (int16_t)(rand() % (2*bound+1)) - bound;
         array[i] = num;
     }
     return array;
 }
 
 // Update position by velocity, one time-step
-void update_coords(int8_t * xs, int8_t * ys, int8_t * zs, int8_t * vx, int8_t * vy, int8_t * vz, int32_t size){
+void update_coords(int16_t * xs, int16_t * ys, int16_t * zs, int16_t * vx, int16_t * vy, int16_t * vz, int32_t size){
     for (int i = 0; i < size; i ++) {
         xs[i] = xs[i] + vx[i];
         ys[i] = ys[i] + vy[i];
@@ -22,8 +22,8 @@ void update_coords(int8_t * xs, int8_t * ys, int8_t * zs, int8_t * vx, int8_t * 
     }
 }
 
-int8_t sum(int8_t * array, int32_t size) {
-    int8_t sum = 0;
+int16_t sum(int16_t * array, int32_t size) {
+    int16_t sum = 0;
     for (int i = 0; i < size; i++) { sum += array[i];}
     return sum;
 }
@@ -37,12 +37,12 @@ int main(int argc, char* argv[]) {
     int32_t size = atoi(argv[1]);
     int32_t iters = atoi(argv[2]);
     srand(size);
-    int8_t * xs = generate_random_list(size, 1000);
-    int8_t * ys = generate_random_list(size, 1000);
-    int8_t * zs = generate_random_list(size, 1000);
-    int8_t * vx = generate_random_list(size, 1);
-    int8_t * vy = generate_random_list(size, 1);
-    int8_t * vz = generate_random_list(size, 1);
+    int16_t * xs = generate_random_list(size, 1000);
+    int16_t * ys = generate_random_list(size, 1000);
+    int16_t * zs = generate_random_list(size, 1000);
+    int16_t * vx = generate_random_list(size, 1);
+    int16_t * vy = generate_random_list(size, 1);
+    int16_t * vz = generate_random_list(size, 1);
 
     struct timespec start, stop;
     clock_gettime(CLOCK_MONOTONIC, &start);
@@ -50,7 +50,7 @@ int main(int argc, char* argv[]) {
         update_coords(xs, ys, zs, vx, vy, vz, size);
     }
     clock_gettime(CLOCK_MONOTONIC, &stop);
-    int8_t chksum = sum(xs, size) + sum(ys, size) + sum(zs, size);
+    int16_t chksum = sum(xs, size) + sum(ys, size) + sum(zs, size);
     double obj_count = (double)size*(double)iters;
     double time_taken;
     time_taken = ((double)stop.tv_sec - (double)start.tv_sec) * 1e9;
