+++ evalAssignment1/update_locations.h
@@ -4,9 +4,9 @@
 #include<stdlib.h>
 #include<time.h>
 
-int8_t * generate_random_list(int32_t size, int32_t bound);
+int16_t * generate_random_list(int32_t size, int32_t bound);
 
-void update_coords(int8_t * xs, int8_t * ys, int8_t * zs, int8_t * vx, int8_t * vy, int8_t * vz, int32_t size);
+void update_coords(int16_t * xs, int16_t * ys, int16_t * zs, int16_t * vx, int16_t * vy, int16_t * vz, int32_t size);
 
-int8_t sum(int8_t * array, int32_t size);
+int16_t sum(int16_t * array, int32_t size);
